feat: relocate getDockerVersion in pkg/driver/docker
